requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
python-dotenv>=1.0.0
pydantic>=2.4.2
sentence-transformers>=2.2.2
chromadb>=0.4.18
google-generativeai>=0.3.0
langchain>=0.0.329
langchain_community>=0.0.10
langchain-text-splitters>=0.0.1
transformers>=4.34.0
torch>=2.0.1
numpy>=1.24.3
python-multipart>=0.0.6
pypdf2>=3.0.0
pdfplumber>=0.10.0
scikit-learn>=1.0.0
tqdm>=4.64.0